\name{cartilage}
\alias{cartilage}
\alias{cartilage}
\title{Spiky cartilage spectra...}
\description{Spiky cartilage spectra
A selection of Raman spectra, the first half with spikes, the second half without}
\docType{data}
\author{A. Bonifacio and C. Beleites}
\keyword{datasets}
\examples{
cartilage
}

